Extract Asana-related helper functions into AsanaHelper
