+++ lib/fastlane/plugin/ddg_apple_automation/actions/asana_extract_task_id_action.rb
@@ -1,13 +1,13 @@
 require "fastlane/action"
 require "fastlane_core/configuration/config_item"
-require_relative "../helper/ddg_apple_automation_helper"
+require_relative "../helper/asana_helper"
 require_relative "../helper/github_actions_helper"
 
 module Fastlane
   module Actions
     class AsanaExtractTaskIdAction < Action
       def self.run(params)
-        Helper::DdgAppleAutomationHelper.extract_asana_task_id(params[:task_url])
+        Helper::AsanaHelper.extract_asana_task_id(params[:task_url])
       end
 
       def self.description
