+++ lib/fastlane/plugin/ddg_apple_automation/actions/asana_get_release_automation_subtask_id_action.rb
@@ -1,6 +1,7 @@
 require "fastlane/action"
 require "fastlane_core/configuration/config_item"
 require "time"
+require_relative "../helper/asana_helper"
 require_relative "../helper/ddg_apple_automation_helper"
 require_relative "../helper/github_actions_helper"
 
@@ -8,7 +9,7 @@ module Fastlane
   module Actions
     class AsanaGetReleaseAutomationSubtaskIdAction < Action
       def self.run(params)
-        Helper::DdgAppleAutomationHelper.get_release_automation_subtask_id(params[:task_url], params[:asana_access_token])
+        Helper::AsanaHelper.get_release_automation_subtask_id(params[:task_url], params[:asana_access_token])
       end
 
       def self.description
