+++ spec/asana_extract_task_assignee_action_spec.rb
@@ -4,7 +4,7 @@ describe Fastlane::Actions::AsanaExtractTaskAssigneeAction do
       task_id = "12345"
       assignee_id = "67890"
       asana_access_token = "secret-token"
-      expect(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:extract_asana_task_assignee)
+      expect(Fastlane::Helper::AsanaHelper).to receive(:extract_asana_task_assignee)
         .with(task_id, asana_access_token).and_return(assignee_id)
       expect(test_action(task_id, asana_access_token)).to eq(assignee_id)
     end
