+++ spec/asana_extract_task_id_action_spec.rb
@@ -1,7 +1,7 @@
 describe Fastlane::Actions::AsanaExtractTaskIdAction do
   describe "#run" do
     it "calls helper" do
-      expect(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:extract_asana_task_id)
+      expect(Fastlane::Helper::AsanaHelper).to receive(:extract_asana_task_id)
         .with("https://app.asana.com/0/0/0").and_return("0")
       expect(test_action("https://app.asana.com/0/0/0")).to eq("0")
     end
