+++ spec/asana_log_message_action_spec.rb
@@ -13,22 +13,22 @@ describe Fastlane::Actions::AsanaLogMessageAction do
       allow(Asana::Client).to receive(:new).and_return(asana_client)
       allow(asana_client).to receive(:tasks).and_return(@asana_client_tasks)
 
-      allow(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:get_release_automation_subtask_id).and_return(automation_subtask_id)
-      allow(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:extract_asana_task_id).and_return(task_id)
-      allow(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:extract_asana_task_assignee).and_return(assignee_id)
-      allow(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:get_asana_user_id_for_github_handle).and_return(assignee_id)
+      allow(Fastlane::Helper::AsanaHelper).to receive(:get_release_automation_subtask_id).and_return(automation_subtask_id)
+      allow(Fastlane::Helper::AsanaHelper).to receive(:extract_asana_task_id).and_return(task_id)
+      allow(Fastlane::Helper::AsanaHelper).to receive(:extract_asana_task_assignee).and_return(assignee_id)
+      allow(Fastlane::Helper::AsanaHelper).to receive(:get_asana_user_id_for_github_handle).and_return(assignee_id)
       allow(@asana_client_tasks).to receive(:add_followers_for_task)
       allow(Fastlane::Actions::AsanaAddCommentAction).to receive(:run)
     end
 
     it "extracts assignee id from release task when is scheduled release" do
-      expect(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:extract_asana_task_id).with(task_url)
-      expect(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:extract_asana_task_assignee).with(task_id, anything)
+      expect(Fastlane::Helper::AsanaHelper).to receive(:extract_asana_task_id).with(task_url)
+      expect(Fastlane::Helper::AsanaHelper).to receive(:extract_asana_task_assignee).with(task_id, anything)
       test_action(task_url: task_url, comment: comment, is_scheduled_release: true)
     end
 
     it "takes assignee id from github handle when is manual release" do
-      expect(Fastlane::Helper::DdgAppleAutomationHelper).to receive(:get_asana_user_id_for_github_handle).with(github_handle)
+      expect(Fastlane::Helper::AsanaHelper).to receive(:get_asana_user_id_for_github_handle).with(github_handle)
       test_action(task_url: task_url, comment: comment, is_scheduled_release: false, github_handle: github_handle)
     end
